# Fix panic when a mesh has skinning attributes but no `SkinnedMesh` component