--- a/Cargo.toml
+++ b/Cargo.toml
@@ -4098,6 +4098,14 @@ description = "Demonstrates specular tints and maps"
 category = "3D Rendering"
 wasm = true
 
+[[example]]
+name = "invalid_skinned_mesh"
+path = "tests/3d/invalid_skinned_mesh.rs"
+doc-scrape-examples = true
+
+[package.metadata.example.invalid_skinned_mesh]
+hidden = true
+
 [profile.wasm-release]
 inherits = "release"
 opt-level = "z"