# Objective

- Fix https://github.com/bevyengine/bevy/issues/16929
- Fix motion blur sometimes being broken on skinned meshes.
- Fix panic when skinned meshes use uniform buffers and motion blur.
- Fix warning spam when a mesh doesn't have skinning attributes but does have a `SkinnedMesh` component (reverse of main bug).
- Add a test that reproduces most of these bugs and is suitable for CI.

## TLDR

Make the renderer consistently decide if a mesh is skinned or unskinned. One of the fixes might be controversial due to performance cost.

## Summary

My goal was to fix https://github.com/bevyengine/bevy/issues/16929 - a wgpu panic when a mesh with skinning attributes is instantiated as a `Mesh3d` but without a `SkinnedMesh` component.

What happened was:

- `setup_morph_and_skinning_defs` checked if the mesh had skinning attributes.
    - The attributes were present, so the pipeline chose a bind group layout *with* skinning.
- `SetMeshBindGroup::render` checked if the mesh was in `skin_uniforms`, which only contains entities with `SkinnedMesh` components.
    - The mesh wasn't in `skin_uniforms`, so `SetMeshBindGroup::render` chose a bind group *without* skinning.
- Now the pipeline didn't match the set bind group, so wgpu panics.

The fix is to make `skin_uniforms` the source of truth.

This was mostly the case already - the exception was `setup_morph_and_skinning_defs`. So I changed the pipeline specialization functions to check `skin_uniforms`, set a new `MeshPipelineKey::SKINNED`, and then pick that up in `setup_morph_and_skinning_defs`.

While working this through I noticed a few more issues.

### Incorrect Pipeline Caches

The pipeline specialization used `RenderMeshInstanceFlags::HAS_PREVIOUS_SKIN` to decide if the mesh needed motion blur. But that flag is set by the `set_mesh_motion_vector_flags` system, which can run *after* specialization, so in some cases skinned meshes had no motion blur.

This is fixed as a side effect of using `skin_uniforms` as the source of truth. As a bonus, this means `RenderMeshInstanceFlags::HAS_PREVIOUS_SKIN` can be removed, and `set_mesh_motion_vector_flags` no longer has to iterate over all skinned entities.

The bad news is that `HAS_PREVIOUS_MORPH` follows a similar pattern, so I suspect motion blur is broken on morph targets in some cases.

### Previous Buffers Too Small

When testing with `skins_use_uniform_buffers` forced on, I got a panic about `SkinUniforms::prev_buffer` being too small for the mapped range. I think the problem is that  `prepare_skins` grows `current_buffer` without also growing `prev_buffer`, so it's too small for one frame.

I've changed `prepare_skins` to grow both buffers. However, that will cause a single frame visual glitch as the new `prev_buffer` is initialised with the *current* frame's joints. For this PR I think reducing a panic to a glitch is sufficient.

I suspect there's also a bug where `prev_buffer` has wrong transforms the first frame after a skin is added. And at the last minute I realised `MeshPipelineKey::HAS_PREVIOUS_SKIN` is probably obsolete, since all skins now have a previous. I might chase these down in a follow-up PR.

### Warning Spam

The original goal was to fix a mesh with skinning attributes but no `SkinnedMesh` component. This is a legitimate use case, e.g. an animated door mesh might be reused as a static door.

I decided to test the opposite as well - what if the mesh has no skinning attributes but does have a `SkinnedMesh` component? This has no legitimate use case as far as I'm aware, but I wanted to make sure the renderer didn't panic. And I kinda just enjoy breaking things.

In practice the renderer didn't panic, but wgpu did spam warnings ("Mesh is missing requested attribute"). This is because `extract_skins` only checks for a `SkinnedMesh` component, and not whether the `Mesh3d`'s asset actually has skinning attributes.

The fix is to make `extract_skins` check the attributes. Unfortunately that adds complication and probably a performance cost (see below), which is annoying when it only fixes a case that's not actually useful. I can drop/separate this change if it's controversial.

There's also an argument that we have now have no warning at all for what's almost certainly a user mistake. Although I'd argue that should be solved by ECS linting rather than the renderer.

### New Test

The PR includes a new test (technically an example) that repros most of the issues. It's designed to be run on CI to detect panics, but can be visually inspected as well. I haven't actually added it to the CI - I wasn't sure whether to do that in this PR or a follow-up.

![image](https://github.com/user-attachments/assets/cdd91bb4-784b-4788-b200-3f82d99b0e55)

### Performance

Tested with `many_foxes`.

The cost of `extract_skins` increases by 10% (178.7us -> 195.9us). *This is not good*. I'm guessing the culprit is the new `Mesh3d` change detection (see "Warning Spam"), but I haven't confirmed that or why it would be so expensive.

On the bright side, the cost of `set_mesh_motion_vector_flags` is much reduced (8.97us -> 0.3us).

## Testing

```
cargo run --example invalid_skinned_mesh
```

Optionally repeat test with `skins_use_uniform_buffers` forced to true.