@alice-i-cecile 2025-02-27
> The fix is to make extract_skins check the attributes. Unfortunately that adds complication and probably a performance cost (see below), which is annoying when it only fixes a case that's not actually useful. I can drop/separate this change if it's controversial.

Let's drop that; the rest of this is a clear improvement that we shouldn't let get bogged down.