@greeble-dev 2025-02-27
I've reverted the attribute checks in `extract_skins`. This removes the performance issue. It also means the `invalid_skinned_mesh` example will now spam warnings, but I'm assuming that's a problem for a follow up PR.

I'll investigate using `SystemParams` tomorrow.