@greeble-dev 2025-02-28
I've [changed](https://github.com/bevyengine/bevy/pull/18074/commits/85d1a00179e96a7380bd2823647940943ec8af6a) the specialize systems to take a `SpecializeParams` struct. This is mainly to dodge the limit on system parameters, but is arguably a bit cleaner too. I could have put more shared params in there but decided to keep things minimal for this PR.

Note that the struct is `pub` because `specialize_material_meshes` is pub. I suspect that's wrong - it should be `pub(crate)` at most? But again, keeping things minimal.